Moved some functions into the cluster_utilities module to avoid importing them from drone_tasks.
